Implemented Nav Bar which routes to pages
