+++ next.config.js
@@ -2,7 +2,6 @@
 const nextConfig = {
   experimental: {
     appDir: true,
-    newNextLinkBehavior: false,
   },
 }
