+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
 import dynamic from "next/dynamic";
 import '../styles/globals.css'
 import '../styles/client.css'
+
 const Navbar = dynamic(
   () => {
     return import('../components/Navbar');
@@ -10,7 +11,7 @@ const Navbar = dynamic(
 
 export const metadata = {
   title: 'Welcome to guevaratorres.space!',
-  description: 'Generated by create next app',
+  description: 'Welcome to Next.js',
 }
 
 export default function RootLayout({
@@ -20,9 +21,9 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      {/* < Navbar />  */}
+      
       <body>
-      < Navbar />
+      < Navbar /> 
       {children}
       </body>
