+++ src/app/page.tsx
@@ -1,28 +1,32 @@
 import Image from 'next/image'
 import { Inter } from 'next/font/google'
-import styles from '../styles/page.module.css'
-import dynamic from "next/dynamic";
 import '../styles/globals.css'
+import Portrait0 from '../components/Project-Images/portrait0.jpg'
 
-const Client = dynamic(
-  () => {
-    return import('./client');
-  },
-  { ssr: false }
-);
-// const Navbar = dynamic(
-//   () => {
-//     return import('../components/Navbar');
-//   },
-//   { ssr: false }
-// );
 const inter = Inter({ subsets: ['latin'] })
 
 export default function page() {
   return (
-    <main className={styles.main}>
-        {/* < Navbar /> */}
-        < Client/>
+    <main>
+      <div>
+        <Image
+          height={500}
+          width={370}
+          src={Portrait0}
+          className="App-port"
+          alt="Portrait"
+        />
+      </div>
+      <div className="About">
+          <h2>About Me</h2>
+        </div>
+
+        <div className="Border">
+          <p>
+            I&#39;m a Software Developer with a Bachelor of Science in Computer
+            Science degree, with an interest in Fullstack applications.
+          </p>
+        </div>
     </main>
-  )
+  );
 }
